Notes:  @someguyinKC had requested the ability to have a misspelled word fixed in your document, if/when you add a new autocorrect item via hh2. That has been added. The feature is meant for scenarios where you're typing some text, then you select a misspelled word (usually by double-clicking it) and activate hh2 via Win+H. If the selected text is multi-line, or if it gets trimmed, then the selected word doesn't get changed. But if you select a misspelled word, then make a whole-word autocorrect entry via hh2, then hh2 will attempt to put the correction into your thing that you are typing. You can disable this, by finding AutoEnterNewEntry := 1 near the top of the code, and setting it to := 0.
This version introduces the ManualCorrectionLogger. By "manual" I mean that it automatically captures and logs manual corrections you make while typing. This is in contrast to the AutoCorrectionLog, with holds the autocorrects that happen via the f() function that is in Autocorrect for v2 (ac2). So the AutoCorrectionLog doesn't need to run in the background. It gets appended, as-needed, by ac2. The ManualCorrectionLogger, on the other hand, is its own application and needs to be running in the background. You'll probably want to put the MCL and put ac2 in your Windows Startup folder. Important: I left my own log data in the logs as an example, and so you can experiment with running the log analyzers, but you'll want to remove my log data and collect your own, that is based on your own typing experiences. My idea with the MCL is that I'd see if particular typos and corrections were reoccurring. If so, I'd consider adding them to my ac2 library of items. I have gotten ideas for new ac items, but that has been from simply manually reviewing the log. I'd added maybe 10 or 15 new items that way. The current number of potential fixes from the ac2 library is just over 360k.

As before, the included exe files are not compiled versions of the scripts. They are copies of AutoHotkey64.exe that have been renamed to match the corresponding .ahk files. The matching ahk and exe files must have the same name, and be in the same folder.